Button toggles directMessagesScreen.
May have a memory leak.  Need to create screen first, and then toggle
the show/hide.  But where to create the screen? In simpleScreen (and
thus in all of the game screens)? in a generic game screen (and then
I need to move the header into the game screen) or in the game mode
(where the code handling the toggle is now - not the best probably)?